chore: rename files to improve project organization
