+++ onboard-app-guilherme/app/login.tsx
@@ -4,9 +4,9 @@ import { StatusBar } from "expo-status-bar";
 import { useState } from "react";
 import { Alert, KeyboardAvoidingView, Platform, SafeAreaView, View } from "react-native";
 import { StyleSheet } from "react-native-unistyles";
-import { FormField } from "../src/components/form-field";
-import { PrimaryButton } from "../src/components/primary-button";
-import { ScreenHeader } from "../src/components/screen-header";
+import { FormField } from "../src/components/form-field/form-field";
+import { PrimaryButton } from "../src/components/primary-button/primary-button";
+import { ScreenHeader } from "../src/components/screen-header/screen-header";
 import { LOGIN_MUTATION } from "../src/graphql/mutations";
 import type { LoginResponse, LoginVariables } from "../src/graphql/types";
 import { saveToken } from "../src/storage/auth";
