+++ onboard-app-guilherme/src/components/screen-header/screen-header-styles.ts
