+++ onboard-app-guilherme/src/components/user-item/user-item-styles.ts
